requirements.txt
crewai
transformer
accelerate
torch
pysqlite3-binary
python-dotenv
langchain-OpenAI
langchain
langchain-community
